- Remove unnecessary files - Remove unnecessary imports - Rename files
