+++ lib/presentation/bloc/global_state/app_global_state_bloc.dart
@@ -1,9 +1,9 @@
-import 'package:bloc/bloc.dart';
 import 'package:flutter_base_rootstrap/data/data_sources/local/abstract/preferences.dart';
 import 'package:flutter_base_rootstrap/presentation/bloc/global_state/app_global_state.dart';
 import 'package:flutter_base_rootstrap/presentation/resources/locale/localize.dart';
 import 'package:flutter_base_rootstrap/presentation/themes/app_themes.dart';
 import 'package:flutter_base_rootstrap/utils/globals.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 
 class AppGlobalStateBloc extends Cubit<AppGlobalState> {
   Preferences get preferences => getIt();
