--- lib/presenter/resources/locale/intl_en.arb
@@ -1,1 +1,0 @@
-{}
